+++ src/semra/web/__init__.py
@@ -1,0 +1,1 @@
+"""Web components."""
