+++ tests/test_web.py
@@ -4,15 +4,24 @@ import unittest
 
 from starlette.testclient import TestClient
 
-from semra import EXACT_MATCH, UNSPECIFIED_MAPPING, Evidence, Mapping, MappingSet, SimpleEvidence
+from semra import (
+    EXACT_MATCH,
+    UNSPECIFIED_MAPPING,
+    Evidence,
+    Mapping,
+    MappingSet,
+    Reference,
+    SimpleEvidence,
+)
 from semra.client import BaseClient, FullSummary, ReferenceHint
 from semra.wsgi import get_app
 from tests.constants import a1, a2, b1, b2
 
+MS1 = MappingSet(
+    name="Test Mapping Set",
+)
 E1 = SimpleEvidence(
-    mapping_set=MappingSet(
-        name="Test Mapping Set",
-    ),
+    mapping_set=MS1,
     justification=UNSPECIFIED_MAPPING,
 )
 M1 = Mapping(subject=a1, predicate=EXACT_MATCH, object=b1, evidence=[E1])
@@ -53,3 +62,31 @@ class TestApp(unittest.TestCase):
 
         res2 = test_client.get(f"/api/evidence/{E1_M2_REFERENCE.curie}")
         self.assertEqual(404, res2.status_code)
+
+        # Test mapping set API
+        res3 = test_client.get(f"/api/mapping_set/{MS1.curie}")
+        self.assertEqual(200, res1.status_code)
+        self.assertEqual(MS1, MappingSet.model_validate(res3.json()))
+
+        res4 = test_client.get("/api/mapping_set/abcdef")
+        self.assertEqual(404, res4.status_code)
+
+        res5 = test_client.get("/api/mapping_set/")
+        self.assertEqual(200, res1.status_code)
+        self.assertEqual([MS1], [MappingSet.model_validate(r) for r in res5.json()])
+
+        # Test mapping API
+        res6 = test_client.get(f"/api/mapping/{M1.curie}")
+        self.assertEqual(200, res1.status_code)
+        self.assertEqual(M1, Mapping.model_validate(res6.json()))
+
+        res7 = test_client.get("/api/mapping/abcdef")
+        self.assertEqual(404, res7.status_code)
+
+        # Test concept API
+        res8 = test_client.get(f"/api/concept/{a1.curie}")
+        self.assertEqual(200, res1.status_code)
+        self.assertEqual(a1, Reference.model_validate(res8.json()))
+
+        res9 = test_client.get("/api/concept/abcasdef")
+        self.assertEqual(404, res9.status_code)
